+++ src/main/java/org/antontech/controller/UserController.java
@@ -21,20 +21,21 @@ public class UserController {
     private UserService userService;
 
     @RequestMapping(value = "", method = RequestMethod.GET)
-    public List<User> getUsers() {
+    public ResponseEntity<List<User>> getUsers() {
         logger.info("I am in getUsers controller");
         List<User> users = userService.getUsers();
-        return users;
+        if(users == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(Collections.emptyList());
+        }
+        return ResponseEntity.ok(users);
     }
 
     @RequestMapping(value = "", method = RequestMethod.POST)
-    public ResponseEntity<String> create(@RequestBody User user) {
+    public String create(@RequestBody User user) {
         logger.info("Post a new user object {}", user.getUserId());
-        boolean isSaved = userService.save(user);
-        if (isSaved)
-            return ResponseEntity.status(HttpStatus.CREATED).body("User is created successfully");
-        else
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Failed to create user");
+        if (userService.save(user))
+            return "User is created successfully";
+        return null;
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
@@ -42,7 +43,7 @@ public class UserController {
         User u = userService.getById(id);
         if (u == null)
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("User with ID " + id + " not found");
-        return ResponseEntity.ok(u);
+        return ResponseEntity.status(HttpStatus.ACCEPTED).body(u);
     }
 
     @RequestMapping(value = "/search/{industry}", method = RequestMethod.GET)
@@ -57,24 +58,36 @@ public class UserController {
     @RequestMapping(value = "/email/{id}", params = {"email"}, method = RequestMethod.PATCH)
     public ResponseEntity updateEmail(@PathVariable(name = "id") long id, @RequestParam(name = "email") String email) {
         logger.info("Pass in variable id: {} and email {}.", id, email);
-        userService.updateEmail(id, email);
-        return ResponseEntity.ok().body("Email for user "+id+" is updated successfully");
+        if(getUserById(id).getStatusCode() == HttpStatus.ACCEPTED) {
+            userService.updateEmail(id, email);
+            return ResponseEntity.ok().body("Email for user " + id + " is updated successfully");
+        } else
+            return  ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("User with ID " + id + " is not found, fail to update the email");
     }
 
     @RequestMapping(value = "/password/{id}", params = {"password"},method = RequestMethod.PATCH)
     public ResponseEntity updatePassword(@PathVariable(name = "id")long id, @RequestParam(name = "password") String password) {
         String encryptedPassword = DigestUtils.md5Hex(password.trim());
         logger.info("Pass in variable id: {} and password {}.", id, encryptedPassword);
-        userService.updatePassword(id, encryptedPassword);
-        return ResponseEntity.ok().body("Password for user " + id + " is updated successfully");
+        if(getUserById(id).getStatusCode() == HttpStatus.ACCEPTED) {
+            userService.updatePassword(id, encryptedPassword);
+            return ResponseEntity.ok().body("Password for user " + id + " is updated successfully");
+        } else
+            return  ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("User with ID " + id + " is not found, fail to create new password");
     }
 
     @RequestMapping(value = "/company/{id}",params = {"company","address","industry"}, method = RequestMethod.PATCH)
     public ResponseEntity updateCompany(@PathVariable(name = "id")long id, @RequestParam(name = "company") String companyName,
                                         @RequestParam(name = "address") String address, @RequestParam(name = "industry") String industry) {
         logger.info("Pass in variable id: {} and company {}.", id, companyName);
-        userService.updateCompany(id, companyName, address, industry);
-        return ResponseEntity.ok().body("Company for user " + id + " is updated successfully");
+        if(getUserById(id).getStatusCode() == HttpStatus.ACCEPTED) {
+            userService.updateCompany(id, companyName, address, industry);
+            return ResponseEntity.ok().body("Company for user " + id + " is updated successfully");
+        } else
+            return  ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("User with ID " + id + " is not found, fail to update company information");
     }
 
     @RequestMapping(value = "/manager/{id}", params = {"firstName","lastName","title","phone"}, method = RequestMethod.PATCH)
@@ -82,15 +95,21 @@ public class UserController {
                                         @RequestParam(name = "lastName")String lastName, @RequestParam(name = "title")String title,
                                         @RequestParam(name = "phone")String phone) {
         logger.info("Pass in variable id: {}, firstNmae {}, lastName{}.", id, firstName,lastName);
-        userService.updateManager(id, firstName, lastName, title, phone);
-        return ResponseEntity.ok().body("Manager for user " + id + " is updated successfully");
+        if(getUserById(id).getStatusCode() == HttpStatus.ACCEPTED) {
+            userService.updateManager(id, firstName, lastName, title, phone);
+            return ResponseEntity.ok().body("Manager for user " + id + " is updated successfully");
+        } else
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("User with ID " + id + " is not found, fail to update manager");
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
     public ResponseEntity delete(@PathVariable(name = "id") long id) {
         logger.info("I am in delete User controller");
-        userService.delete(id);
-        return ResponseEntity.ok().body("User " + id + " is deleted successfully");
+        if(getUserById(id).getStatusCode() == HttpStatus.ACCEPTED) {
+            userService.delete(id);
+            return ResponseEntity.ok().body("User " + id + " is deleted successfully");
+        } else
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("User with ID " + id + " is not found, fail to delete.");
     }
 
 }
